requirements.txt
pytz  # https://github.com/stub42/pytz
python-slugify  # https://github.com/un33k/python-slugify
Pillow # https://github.com/python-pillow/Pillow
argon2-cffi  # https://github.com/hynek/argon2_cffi
whitenoise  # https://github.com/evansd/whitenoise
redis # https://github.com/andymccurdy/redis-py
celery # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat # https://github.com/celery/django-celery-beat

# Django
# ------------------------------------------------------------------------------
django   # https://www.djangoproject.com/
django-environ  # https://github.com/joke2k/django-environ
django-model-utils  # https://github.com/jazzband/django-model-utils
django-allauth  # https://github.com/pennersr/django-allauth
django-crispy-forms  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis # https://github.com/niwinz/django-redis
# Django REST Framework
djangorestframework  # https://github.com/encode/django-rest-framework


Werkzeug # https://github.com/pallets/werkzeug
ipdb  # https://github.com/gotcha/ipdb
Sphinx # https://github.com/sphinx-doc/sphinx
psycopg2-binary  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
mypy  # https://github.com/python/mypy
django-stubs  # https://github.com/typeddjango/django-stubs
pytest # https://github.com/pytest-dev/pytest
pytest-sugar  # https://github.com/Frozenball/pytest-sugar

# Code quality
# ------------------------------------------------------------------------------
flake8  # https://github.com/PyCQA/flake8
flake8-isort  # https://github.com/gforcada/flake8-isort
coverage  # https://github.com/nedbat/coveragepy
black  # https://github.com/ambv/black
pylint-django  # https://github.com/PyCQA/pylint-django
pylint-celery  # https://github.com/PyCQA/pylint-celery
pre-commit  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar  # https://github.com/jazzband/django-debug-toolbar
django-extensions  # https://github.com/django-extensions/django-extensions
django-coverage-plugin  # https://github.com/nedbat/django_coverage_plugin
pytest-django # https://github.com/pytest-dev/pytest-django


gunicorn  # https://github.com/benoitc/gunicorn
# psycopg2==2.8.4 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
sentry-sdk  # https://github.com/getsentry/sentry-python

# Django
# ------------------------------------------------------------------------------
django-storages[boto3]  # https://github.com/jschneier/django-storages
django-anymail[mailgun]  # https://github.com/anymail/django-anymail

# Extra
# ------------------------------------------------------------------------------
tweepy
django-grappelli
